+++ src/main/java/qna/repository/UserRepository.java
@@ -1,6 +1,7 @@
-package qna.domain;
+package qna.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.User;
 
 import java.util.Optional;
